+++ NEWS.md
@@ -1,6 +1,7 @@
 ### Unreleased
 
 * Scripting: Added -e,--evaluate to run a script from command-line
+* Split up object types file type selection
 * Raised minimum supported Qt version from 5.6 to 5.12 (drops Windows XP support)
 * Raised minimum C++ version to C++17
 * Removed qmake project files (only Qbs supported now)